+++ next-sitemap.config.js
@@ -1,7 +1,0 @@
-/** @type {import('next-sitemap').IConfig} */
-
-module.exports = {
-  siteUrl: 'https://shlink.io',
-  generateRobotsTxt: true,
-  outDir: 'public'
-}