+++ src/components/Breadcrumb.tsx
@@ -1,19 +1,22 @@
-import type { FC, PropsWithChildren } from 'react';
+import type { FC } from 'react';
 import { Breadcrumb as BsBreadcrumb, BreadcrumbItem } from 'reactstrap';
-import type { Route } from '../utils/docUtils';
+import { useBreadcrumbsForPath } from '../utils/docUtils';
 import { Link } from './Link';
 
-type BreadcrumbProps = PropsWithChildren<{
-  prevElements: Route[];
-}>;
+type BreadcrumbProps = {
+  currentPage: string;
+};
 
-export const Breadcrumb: FC<BreadcrumbProps> = ({ children, prevElements }) => (
-  <BsBreadcrumb>
-    {prevElements.map(({ text, link }) => (
-      <BreadcrumbItem key={text}>
-        <Link href={link}>{text}</Link>
-      </BreadcrumbItem>
-    ))}
-    <BreadcrumbItem active>{children}</BreadcrumbItem>
-  </BsBreadcrumb>
-);
+export const Breadcrumb: FC<BreadcrumbProps> = ({ currentPage }) => {
+  const { breadcrumbItems, title } = useBreadcrumbsForPath(currentPage);
+  return (
+    <BsBreadcrumb>
+      {breadcrumbItems.map(({ text, link }) => (
+        <BreadcrumbItem key={text}>
+          <Link href={link}>{text}</Link>
+        </BreadcrumbItem>
+      ))}
+      <BreadcrumbItem active>{title}</BreadcrumbItem>
+    </BsBreadcrumb>
+  );
+};