+++ src/components/InternalLink.tsx
@@ -1,5 +1,4 @@
-import Link from 'next/link';
 import type { FC } from 'react';
 import type { LinkProps } from './Link';
 
-export const InternalLink: FC<LinkProps> = ({ children, ...rest }) => <Link {...rest}>{children}</Link>;
+export const InternalLink: FC<LinkProps> = ({ children, ...rest }) => <a {...rest}>{children}</a>;