+++ src/utils/docUtils.ts
@@ -8,6 +8,7 @@ import {
   faShieldAlt,
   faTerminal,
 } from '@fortawesome/free-solid-svg-icons';
+import { useMemo } from 'react';
 
 export interface Route {
   text: string;
@@ -149,5 +150,11 @@ export const breadcrumbsMap: BreadcrumbsMap = docsRoutes.reduce<BreadcrumbsMap>(
   {},
 );
 
-export const breadcrumbForPath = (path: string): BreadcrumbInfo =>
-  breadcrumbsMap[path] || { title: '', breadcrumbItems: [] };
+export const normalizePath = (path: string) => (path.endsWith('/') ? path.substring(0, path.length - 1) : path);
+
+export const breadcrumbForPath = (path: string): BreadcrumbInfo => {
+  const normalizedPath = normalizePath(path);
+  return breadcrumbsMap[normalizedPath] || { title: '', breadcrumbItems: [] };
+};
+
+export const useBreadcrumbsForPath = (path: string) => useMemo(() => breadcrumbForPath(path), [path]);