+++ src/utils/markdownUtils.tsx
@@ -3,31 +3,10 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import type { MDXComponents } from 'mdx/types';
 import { Mermaid } from 'mdx-mermaid/lib/Mermaid';
 import type { PropsWithChildren } from 'react';
-import Highlight from 'react-highlight';
 import { Link } from '../components/Link';
 
-const slugify = (str: string) => {
-  let lowercasedStr = str
-    .replace(/^\s+|\s+$/g, '') // trim
-    .toLowerCase();
-
-  // remove accents, swap ñ for n, etc
-  const from = 'åàáãäâèéëêìíïîòóöôùúüûñç·/_,:;';
-  const to = 'aaaaaaeeeeiiiioooouuuunc------';
-
-  for (let i = 0, l = from.length; i < l; i += 1) {
-    lowercasedStr = lowercasedStr.replace(new RegExp(from.charAt(i), 'g'), to.charAt(i));
-  }
-
-  return lowercasedStr
-    .replace(/[^a-z0-9 -]/g, '') // remove invalid chars
-    .replace(/\s+/g, '-') // collapse whitespace and replace by -
-    .replace(/-+/g, '-'); // collapse dashes
-};
-
 const buildTitleForTag = (Tag: string) => (props: PropsWithChildren<any>) => { // eslint-disable-line
-  const { children, ...rest } = props;
-  const link = slugify(children);
+  const { children, id: link, ...rest } = props;
 
   return (
     <Tag {...rest}>
@@ -40,23 +19,16 @@ const buildTitleForTag = (Tag: string) => (props: PropsWithChildren<any>) => { /
   );
 };
 
-const a = (props: any) => <Link {...props} />;
-const pre = (props: any) => {
-  if (props.children?.props?.className?.includes('mermaid')) {
-    return <Mermaid chart={props.children?.props?.children} />;
-  }
-
-  return <Highlight {...props} />;
-};
 const h1 = buildTitleForTag('h1');
 const h2 = buildTitleForTag('h2');
 const h3 = buildTitleForTag('h3');
 const h4 = buildTitleForTag('h4');
 
+const a = (props: any) => <Link {...props} />;
 const table = (props: any) => (
   <div className="table-responsive">
     <table className="table table-bordered" {...props} />
   </div>
 );
 
-export const markdownComponents: MDXComponents = { a, pre, h1, h2, h3, h4, table };
+export const markdownComponents: MDXComponents = { h1, h2, h3, h4, a, table, mermaid: Mermaid, Mermaid };